pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "metahub"
description = ''
readme = "README.md"
requires-python = ">=3.7"
license = "MIT"
keywords = []
authors = [
  { name = "Xavier Petit", email = "nuxion@gmail.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
	"Django~=4.2.3",
	"djangorestframework~=3.14.0",
        "gunicorn~=20.1.0",
        "drf-spectacular~=0.26.3",
]
dynamic = ["version"]

[project.optional-dependencies]
jupyter = [
   "jupyterlab~=3.6.2",
   "jupytext~=1.14.5",
]


[project.urls]
Documentation = "https://github.com/nuxion/metahub#readme"
Issues = "https://github.com/nuxion/metahub/issues"
Source = "https://github.com/nuxion/metahub"

[tool.hatch.version]
path = "metahub/__about__.py"


[tool.hatch.envs.default]
dependencies = [
  "pytest",
  "pytest-asyncio",
  "pytest-cov",
  "pylint~=2.16.2",
  "isort~=5.12.0",
  "black~=23.1.0",
  "black-macchiato",  # for partial update of buffers in emacs
  "mypy~=1.0.1",
  "Sphinx~=5.3.0",
  "sphinx-autobuild",
  "sphinx-rtd-theme",
  "httpx",
  "ipython",
  "django-stubs",
  "pip-tools"
]


[tool.hatch.envs.jupyter]
features = ["jupyter"]

[tool.hatch.envs.default.scripts]
cov = "pytest --cov-report=term-missing --cov-config=pyproject.toml --cov=metahub --cov=tests {args}"
no-cov = "cov --no-cov {args}"

[[tool.hatch.envs.test.matrix]]
python = ["37", "38", "39", "310", "311"]

[tool.coverage.run]
branch = true
parallel = true
omit = [
  "metahub/__about__.py",
]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

[tool.isort]
profile = "black"


[tool.black]
line-length = 88

[tool.jupytext]
formats = "ipynb,py"
